Prevent windows platform (GOOS) from using voidstar_handler
