+++ internal/sdk_const.go
@@ -3,7 +3,7 @@ package internal
 // --------------------------------------------------------------------------------
 // Version
 // --------------------------------------------------------------------------------
-const SDK_Version = "0.3.8"
+const SDK_Version = "0.3.9"
 
 // --------------------------------------------------------------------------------
 // Environment Vars
